+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * more time-related calls will accept time strings
 * added optional `delay` argument to `call_in_separate_thread`
+* beefed up BogusTerminableThread docs